Added code to prune subgraph during path search
I discovered that we can iteratively prune the subgraph
extracted for path search. We do the double sided search to get edges
for the initial subgraph. Then we use BFS from X and from Y and take intersection,
pruning all edges that lie outside intersection. Now we repeat this until the graph
stops changing. This can help during the exhaustive search of all paths, since the
GraphSearchTree is not that intelligent.