Fixed PathGraph Bug
Fixed a bug where makePathGraph would return an empty graph even when
a path existed. This was due to faulty logic which said return empty graph if
X edges to halfDistanceX is empty, or if Y edges to halfDistanceY is empty.
It is possible for one of these edge lists to be empty and for a path to still exist
satisfying the path length constraints.