Fixed parallelization bug
There was a bug in Subgraph::copyEdgesToSubgraph. This function
used to hold a copy of the main StringGraph and modified it by marking
the edges that had been already added to subgraph (to avoid adding duplicate
edges). However, all threads would be modifiying the main graph in this way,
causing non-deterministric results. Each thread should be using a const version of the
main graph.