Added Vertex getEdgesBegin()/End()
Added getEdgesBegin() and getEdgesEnd() to provide
fast access to a list of vertex edges. This avoids populating
a vector of edges. It could potentially be used anywhere
the list of Vertex edges needs to be iterated over, without
changing the list itself.

I updated the GraphSearchTree and code in BFS/dijkstra's to make
use of this. sga close-path runs about 25% faster in single threaded
mode with this change.