SGWalkVector copy reduction
The SGWalkVector is a Vector of SGWalk instances, which contain a
vector of edges.

Therefore, dealing with SGWalkVector's can incur many reallocations
for vectors as the SGWalkVector itself grows or as a copy of the SGWalkVector
is made.

I reduced the need to copy SGWalkVector's through the use of std::vector.swap(),
and I made modifications to the GraphSearchTree and WalkBuilder to allow more efficient
construction of walks.